Modified code to avoid duplication

+++ vitals/Battery.java
@@ -1,25 +1,24 @@
 package vitals;
 
 /**
- * @author PDH2COB
  * Class with battery parameters
  */
 public class Battery {
 
-	private final int temperature;
+	private final float temperature;
 
 	private final float chargeRate;
 
-	private final int soc;
+	private final float soc;
 
-	public Battery(final int temperature, final float chargeRate, final int soc) {
+	public Battery(final float temperature, final float chargeRate, final float soc) {
 		super();
 		this.temperature = temperature;
 		this.chargeRate = chargeRate;
 		this.soc = soc;
 	}
 
-	public int getTemperature() {
+	public float getTemperature() {
 		return temperature;
 	}
 
@@ -27,8 +26,10 @@ public class Battery {
 		return chargeRate;
 	}
 
-	public int getSoc() {
+	public float getSoc() {
 		return soc;
 	}
+
+
 
 }
